psm/requirements.txt
asgiref==3.5.0
attrs==22.1.0
backports.zoneinfo==0.2.1
cached-property==1.5.2
certifi==2022.6.15
cffi==1.15.0
charset-normalizer==2.1.1
click==8.1.2
cryptography==37.0.2
dj-database-url==0.5.0
Django==4.0.3
django-better-admin-arrayfield==1.4.2
django-extensions==3.1.5
django-heroku==0.3.1
django-json-widget==1.1.1
djangorestframework==3.13.1
future==0.18.2
gunicorn==20.1.0
h11==0.13.0
idna==3.3
isodate==0.6.1
lxml==4.9.1
platformdirs==2.5.2
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
uvicorn==0.17.6
whitenoise==6.0.0
XlsxWriter==3.0.3
zeep==4.1.0
